+++ plugins/fyp/src/components/Sarah-Component/Metrics/MeanTimeToRecovery.tsx
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2023 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import React, { Component } from 'react';
+import { Typography } from '@material-ui/core';
+
+class MeanTimeToRecovery extends Component {
+  render() {
+    return (
+      <div className="container mt-5">
+        <h2>Metric 3: Mean Time to Recovery</h2>
+
+        <Typography>Mean Time to Recovery:</Typography>
+        <Typography>Performance Result: </Typography>
+      </div>
+    );
+  }
+}
+
+export default MeanTimeToRecovery;
